Change regex pattern and move test under unit/command
